Refactoring UI into discrete types
