+++ src/portal/ui-binding.ts
@@ -2,6 +2,9 @@
 // This is primarily done to enable testing.
 
 import * as api from './api-client';
+import RemediationGrid from './remediation-grid';
+import ScheduledJobsGrid from './scheduled-jobs-grid';
+import { UiCommon } from './ui-common';
 
 export interface IUIBinding {
     // Global event hook setup + busy / spinner mask control
@@ -9,6 +12,7 @@ export interface IUIBinding {
         onDialogClose: () => void,
         onRefreshContent: () => void
     ): void;
+
     SetDetailsLinkCallback(onDetailsClick: (id: string) => void): void;
     SetBusyState(busy: boolean): void;
 
@@ -140,156 +144,20 @@ export class UIBinding implements IUIBinding {
     }
 
     public RenderRemediationNeededContent(data: api.IRemediationNeeded) {
-        const markup = `
-            <div class="dataRegion">
-                <p>The following systems are found to need compute upgrades:</p>
-                <div class="dataGridToolbar">
-                    <ul class="listNone">
-                        <li title="Group Selected"><i class="far fa-object-group"></i>Group Selected</li>
-                        <li title="Ungroup Selected"><i class="far fa-object-ungroup"></i>Ungroup Selected</li>
-                        <li class="refreshContent" title="Refresh"><i class="fas fa-sync"></i>Refresh</li>
-                    </ul>
-                </div>
-                <table class="dataGrid">
-                    <colgroup>
-                        <col width="25px" />
-                        <col width="200px" />
-                        <col width="100px" />
-                        <col width="200px" />
-                        <col width="200px" />
-                        <col width="*" />
-                    </colgroup>
-                    <tr class="headerRow">
-                        <td>&nbsp;</td>
-                        <td>Name</td>
-                        <td>Group</td>
-                        <td>Type</td>
-                        <td>Description</td>
-                        <td>Details</td>
-                    </tr>
-                    ${this.renderRemediationNeededRows(data)}
-                </table>
-            </div>
-            <div class="dataRegion">
-                <p>The following systems are found to need storage account upgrades:</p>
-                <div class="dataGridToolbar">
-                    <ul class="listNone">
-                        <li title="Group Selected"><i class="far fa-object-group"></i>Group Selected</li>
-                        <li title="Ungroup Selected"><i class="far fa-object-ungroup"></i>Ungroup Selected</li>
-                        <li class="refreshContent" title="Refresh"><i class="fas fa-sync"></i>Refresh</li>
-                    </ul>
-                </div>
-                <table class="dataGrid">
-                    <colgroup>
-                        <col width="25px" />
-                        <col width="200px" />
-                        <col width="100px" />
-                        <col width="200px" />
-                        <col width="200px" />
-                        <col width="*" />
-                    </colgroup>
-                    <tr class="headerRow">
-                        <td>&nbsp;</td>
-                        <td>Name</td>
-                        <td>Group</td>
-                        <td>Type</td>
-                        <td>Description</td>
-                        <td>Details</td>
-                    </tr>
-                    ${this.renderRemediationNeededRows(data)}
-                </table>
-            </div>
-        `;
-
+        const g = new RemediationGrid();
+        const markup = g.RenderNeeded(data);
         $('.content-stage .placeholder').html(markup);
     }
 
     public RenderRemediationCompletedContent(data: api.IRemediationCompleted) {
-        const markup = `
-            <div class="dataRegion">
-                <p></i>The following systems have completed compute ugprades:</p>
-                <div class="dataGridToolbar">
-                    <ul class="listNone">
-                        <li class="refreshContent" title="Refresh"><i class="fas fa-sync"></i>Refresh</li>
-                    </ul>
-                </div>
-                <table class="dataGrid">
-                    <colgroup>
-                        <col width="200px" />
-                        <col width="100px" />
-                        <col width="200px" />
-                        <col width="100px" />
-                        <col width="*" />
-                    </colgroup>
-                    <tr class="headerRow">
-                        <td>Name</td>
-                        <td>Group</td>
-                        <td>Type</td>
-                        <td>Duration</td>
-                        <td>Details</td>
-                    </tr>
-                    ${this.renderRemediationCompletedRows(data)}
-                </table>
-            </div>
-            <div class="dataRegion">
-                <p>The following systems have completed storage account upgrades:</p>
-                <div class="dataGridToolbar">
-                    <ul class="listNone">
-                        <li class="refreshContent" title="Refresh"><i class="fas fa-sync"></i>Refresh</li>
-                    </ul>
-                </div>
-                <table class="dataGrid">
-                    <colgroup>
-                        <col width="200px" />
-                        <col width="100px" />
-                        <col width="200px" />
-                        <col width="100px" />
-                        <col width="*" />
-                    </colgroup>
-                    <tr class="headerRow">
-                        <td>Name</td>
-                        <td>Group</td>
-                        <td>Type</td>
-                        <td>Duration</td>
-                        <td>Details</td>
-                    </tr>
-                    ${this.renderRemediationCompletedRows(data)}
-                </table>
-            </div>
-        `;
-
+        const g = new RemediationGrid();
+        const markup = g.RenderCompleted(data);
         $('.content-stage .placeholder').html(markup);
     }
 
     public RenderScheduledJobsContent(data: api.IScheduledJobs) {
-        const markup = `
-            <div class="dataRegion">
-                <p></i>Current scheduled job status:</p>
-                <div class="dataGridToolbar">
-                    <ul class="listNone">
-                        <li class="refreshContent" title="Refresh"><i class="fas fa-sync"></i>Refresh</li>
-                    </ul>
-                </div>
-                <table class="dataGrid">
-                    <colgroup>
-                        <col width="200px" />
-                        <col width="200px" />
-                        <col width="100px" />
-                        <col width="100px" />
-                        <col width="*" />
-                    </colgroup>
-                    <tr class="headerRow">
-                        <td>Job Type</td>
-                        <td>Status</td>
-                        <td>Duration</td>
-                        <td>Job Log</td>
-                        <td>Last Update</td>
-                    </tr>
-                    ${this.renderScheduledJobsRows(data)}
-                </table>
-            </div>
-        `;
-
+        const g = new ScheduledJobsGrid();
+        const markup = g.Render(data);
         $('.content-stage .placeholder').html(markup);
     }
 
@@ -298,7 +166,7 @@ export class UIBinding implements IUIBinding {
             <h2>Remediation Details</h2>
             <ul class="listNone marginBottom">
                 <li><strong>Name:</strong> ${data.Name} (${data.Type})</li>
-                <li><strong>Duration:</strong> ${this.formatDurationInMs(
+                <li><strong>Duration:</strong> ${UiCommon.FormatDurationInMs(
                     data.DurationInMs
                 )}</li>
             </ul>
@@ -307,59 +175,5 @@ export class UIBinding implements IUIBinding {
 
         $('.dialog-stage .placeholder').html(markup);
         $('.dialog-stage').show();
-    }
-
-    private formatDurationInMs(durationInMs: number) {
-        const mins = durationInMs / (60 * 1000);
-        return `${mins.toFixed(2)} mins`;
-    }
-
-    private renderRemediationNeededRows(data: api.IRemediationNeeded) {
-        return data.NeedsComputeUpgrade.map(item => {
-            return `
-            <tr>
-                <td><input type="checkbox" /></td>
-                <td>${item.Name}</td>
-                <td>${item.Group}</td>
-                <td>${item.Type}</td>
-                <td>${item.UpgradeDescription}</td>
-                <td><a class="detailsViewLink" data-item-name="${
-                    item.Name
-                }">Click to view...</a></td>
-            </tr>
-            `;
-        }).join('');
-    }
-
-    private renderRemediationCompletedRows(data: api.IRemediationCompleted) {
-        return data.HadComputeUpgraded.map(item => {
-            return `
-            <tr>
-                <td>${item.Name}</td>
-                <td>${item.Group}</td>
-                <td>${item.Type}</td>
-                <td>${this.formatDurationInMs(item.DurationInMs)}</td>
-                <td><a class="detailsViewLink" data-item-name="${
-                    item.Name
-                }">Click to view...</a></td>
-            </tr>
-            `;
-        }).join('');
-    }
-
-    private renderScheduledJobsRows(data: api.IScheduledJobs) {
-        return data.JobList.map(item => {
-            return `
-            <tr>
-                <td>${item.JobType}</td>
-                <td>${item.Status}</td>
-                <td>${this.formatDurationInMs(item.DurationInMs)}</td>
-                <td><a class="detailsViewLink" data-item-name="${
-                    item.JobId
-                }">View Log...</a></td>
-                <td>${item.LastUpdate.toLocaleDateString()} ${item.LastUpdate.toLocaleTimeString()}</td>
-            </tr>
-            `;
-        }).join('');
     }
 }
